LU-8712 tests: refresh mgs config database before add mds
conf-sanity test 72 failed to start mds because mds failed
with -EADDRINUSE. Registering mdt to MGS while mounting mds
index of mdt was already present in MGS configuration
database so mds mount failed with -EADDRINUSE

Fix:
call reformat to format the mgs when mgs and mds is separate.
This will make sure index of mdt will not be present in MGS
config database.

Signed-off-by: Jadhav Vikram <jadhav.vikram@seagate.com>
Change-Id: Idb317b4c57a902efb269afae968048367b656d1b