LU-4939 utils: Add scripts to output MGS config parameters
Add two bash scripts to read and output OST pools and lctl set_param
parameters.  These scripts will remove any history and output only
the current active parameters.  If a pool is created and destroyed
nothing will be displayed for that pool.  If a parameter is changed
multiple times, only the most recent setting is displayed.

These scripts are helpful in migrating configurations, or backing
up and restoring configurations during a writeconf.

Signed-off-by: Ben Evans <bevans@cray.com>
Change-Id: I7ceffea2eddecc0585d075b61cab8049616817ae