LU-8709 llite: implement parallel asynchronous readahead
This patch proposes a new design of Lustre readahead framework,
parallel-readahead, which implements a similar algorithm like current
Linux kernel's on-demand readahead, but in a way that is more
optimized for the characteristics of Lustre, more extendable for
differet I/O patterns, and more adaptable to the high bandwidth of
Lustre since data could be prefetched in parallel by multiple
threads.

Like the on-demand readahead, the parallel-readahead maintains a
"readahead window" consisting of a portion of the file starting
with the application's last read. Pages inside the readahead window
are either already in the page cache, or being prefetched into the
page cache of the client. The window moves forward as the reading of
the application moves forward.

The parallel-readahead framework prefetches more agressively than
on-demand readahead using a large readahead window, since Lustre
usually has more overhead than local file systems.

The parallel-readahead can do real asynchronous prefetch in parallel
by multiple threads. The reason of this implement is that, if all
prefetching is done in the process of the application that issued
read() syscall, the overhead caused by logic of Lustre client side
would become the bottleneck.

The parallel-readahead framework provides an extendable I/O pattern
detection mechanism which currently has detection algorithms for both
sequential read and stride read.  And more pattern detection
algorithms and corresponding prefetching policies can be added into
the framework for other I/O patterns, including random read,
semisequential read, backward read, interleaved read, etc.

Test-Parameters: fortestonly
Signed-off-by: Li Xi <lixi@ddn.com>
Change-Id: Ida399b9d0b499ba3f6bd08d7bb4528132e8e98e4