LU-8812 test: add failback test with errors=panic
This is test for MRP-1658 which create activity and
emulate failback operations. Option errors=panic
is using similar to production servers. Test requires
ENABLE_QUOTA=yes option.

Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/1550

SEA-107 tests: terminal IO process in end

It is unnecessary to wait IO process in test_92. It
may spend a long time to finish on a big OST.

Signed-off-by: Alexander Boyko <alexander_boyko@xyratex.com>
Change-Id: Iae6765b747ff266e6a44df8ebcd48cfba1991837
Tested-by: Elena Gryaznova <elena_gryaznova@xyratex.com>
Reviewed-by: Elena Gryaznova <elena_gryaznova@xyratex.com>
Reviewed-by: Andrew Perepechko <andrew_perepechko@xyratex.com>
Reviewed-by: Vitaly Fertman <vitaly_fertman@xyratex.com>
Signed-off-by: Yang Sheng <yang.sheng@intel.com>