LU-4423 lnet: avoid unnecessary timeval conversion
The lnet_eq_wait_locked tries to wait for time to pass or an event to
wake up the wait queue. The entire logic seems to be a very elaborate
reimplementation of wait_event().

I'm not trying to clean up the entire logic here, but this at least
gets rid of the multi-way conversion between miliseconds, timeval
and jiffies.

Only lnet_ping passes a real timeout value to lnet_eq_wait_locked.
Since this is the case we can turn the user passed in time out value
into a time64_t to be used by lnet_eq_wait_locked. The reason we don't
pass in jiffies directly is that it is a unsigned long which would
not work with the if (tm < 0) code.

Linux-commit: a11ef8ca94cccca98e0d7f23d5145cad71dd62f0

Change-Id: I15126aab500e975ec34ee270d3a1f9b83cf50902
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Oleg Drokin <green@linuxhacker.ru>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>