LU-8272 ldlm: Use interval tree to update kms
Currently, ldlm_extent_shift_kms does a linear search of
the list of granted locks on the resource when looking for
the lock to use to update the kms.

When cleaning up all locks on a resource, we end up doing
a linear search of this list for every lock on the object;
so, complexity n^2.

For resources with thousands of locks, this can take
multiple seconds, and scales quickly to tens of seconds per
resource.  This can happen in normal usage with a weird I/O
pattern, but it is particularly a problem with lock ahead,
which creates large numbers of locks deliberately.

This can be avoided by using the interval trees which store
the extents of granted locks.  For PW/write locks, the lock
with the highest start must be the lock with the highest
end as well, so we can walk the interval tree in reverse to
almost immediately find the new 'highest end'.

Since the tree is sorted by 'start' and PR locks can
overlap, we cannot easily use the tree to find the PR lock
with the 'highest end'.  But we can still dramatically
improve the average performance by walking the tree from
the highest starting lock, since locks with a high starting
point will more often have high end points.

Note that we test in_max_high for each tree, and if kms is
> in_max_high, we skip that tree.  This is safe because the
kms_ignore flag means that kms from that tree can be lower
than in_max_high for that tree, but it can never be higher,
since in_max_high includes locks with the kms_ignore flag
set.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I9efa4733e691cb2299049ba917325b939be52069