LU-6179 llite: Implement ladvise requestlock
Ladvise requestlock is a new feature allowing userspace to
request extent locks in advance of the IO which will use
them. These locks are not expanded beyond the size
requested by userspace.  They are intended to make it
possible to address lock contention between multiple
clients resulting from lock expansion.  They should allow
optimizing various IO patterns, notably strided writing.
(Further information in LU-6179)

Asynchronous glimpse locks are a speculative version of
glimpse locks, and already implement the required behavior.
Therefore, this patch takes the CEF_AGL flag and renames it
to CEF_SPECULATIVE.

There is no change to the behavior the flag invokes, but it
is now shared between AGL and requestlock lock requests.

Additionally, requestlock creates extent locks in advance
of IO, and so breaks the assumption that the holder of the
highest lock knows the current file size.

So we also modify the ofd_intent_policy code to glimpse
PW locks until it finds one it knows to be in use, taking
care to send only one glimpse to each client.

The current patch allows only non-blocking lock requests,
this avoids deadlocks that would otherwise occur in having
ptlrpcd threads handle blocking lock requests.

LDLM_FL_BLOCK_NOWAIT was only implemented for group locks,
so this patch also makes the changes necessary to allow it
for these manually requested locks.

Finally, this patch also adds another advice to disable
lock expansion, setting a per-file descriptor flag.  This
allows user space to control whether or not lock requests
on this file descriptor will undergo lock expansion.

This means if requestlock locks are not created ahead of IO
(due to inherent raciness) or are cancelled by a competing
IO request, the IO requests that should have used the
manually requested locks will not result in expanded locks.
This avoids lock ping-pong, and because the resulting locks
will not extend to the end of the file, future requestlock
requests can be granted.  Effectively, this means that if
requestlock usage for strided IO is interrupted by a
competing request, it can re-assert itself.

Requestlock is implented via the ladvise interface from
userspace.  As requestlock results in a DLM lock request
rather than file advice, we do not use the lower levels of
the ladvise implementation.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I1e80286f54946a0df08b19b1339829fcfd1117e7