LU-8752 lnet: Debugging Patch
This is a debugging patch only.  Do not review!!

Thanks to weak module loading, I have lost the ability to
load changes to lnet.ko on our test cluster nodes for
debugging.  I now have to submit debugging to Gerrit
to make it into a complete Lustre build which can be
loaded onto a test cluster node.  I could not envision
a less efficient way to be doing this.

Test-Parameters: trivial
Signed-off-by: Doug Oucharek <doug.s.oucharek@intel.com>
Change-Id: If57c21e4dfc1b6d3f6d37b3b46f06eda5e0a73f4