LU-8911 tests: sanity-hsm test_24d fails on a local setup
In test_24d, do not use the default mountpoint of copytool_setup()
as it is the alose the one the test mounts in read-only mode.

This patch also removes the "continuing fast path" in copytool_setup
(calling "pkill -CONT -x lhsmtool_posix" instead of launching a new
 copytool) as it is not explicitly used by any test as well as the
first copytool_setup located outside of any tests.

Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Change-Id: I86bc41213bc656d7a83d63fb8e9bc595ba6b73ca