LU-8936 llite: use percpu env correctly in ll_invalidatepage
This issue is introduced by LU-8509.

In LU-4257, a patch(commit 4533271) is landed to get rid of cl_env hash
table on the client side. From now on, cl_env won't be attached to
the running processes. This made it possible to use cl_env_percpu_get() 
in ll_invalidatepage() as in LU-8509.

However, in LLNL FE branch it includes LU-8509 but doesn't have the
corresponding fix of commit 4533271, which caused the problem on this
ticket because an env is already attached to the process when it
calls ll_invalidatepage().

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: Ie2a2d8276abf929d90eb3a64d72c0d9c911bf2f2