LU-8912 nodemap: fix contiguous range support
This patch fixes the contiguous range check to allow the addition of
multiple "full" ([0-255]) ranges. As part of this change,
is_contiguous and find_min_max are combined as they were always
called together and the logic is fairly similar. This also removes
the multiple range expression support, since it was broken.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I3c49a077039327fcbde87196f82db140f67a74d0