LU-6245 lnet: pass jiffies to lnet_eq_wait_locked
The reason we don't pass in jiffies directly is that it is a
unsigned long which would not work with the if (tm < 0) code.
We can simplify the code by handling the less than zero testing
directly on the timestamp passed in by user land. Then change
that millisecond timestamp to a jiffy. If the timestamp is
negative just set jiffies to MAX_JIFFY_OFFSET.

Change-Id: I15126aab500e975ec34ee270d3a1f9b83cf50902
Signed-off-by: James Simmons <uja.ornl@yahoo.com>